makeing use of our 3 new buffers

+++ compiler/src/jdk.graal.compiler/src/jdk/graal/compiler/hotspot/meta/Bubo/BuboCache.java
@@ -10,7 +10,7 @@ import java.lang.Thread;
  public class BuboCache extends Thread {
         
         public static long[] TimeBuffer;
-        public static int[] ActivationCountBuffer;
+        public static long[] ActivationCountBuffer;
         public static long[] CyclesBuffer;
         public static int pointer;
         public static long[] Buffer;
@@ -18,7 +18,7 @@ import java.lang.Thread;
         public BuboCache() {
                 Buffer = new long[200_000];
                 TimeBuffer = new long[200_000];
-                ActivationCountBuffer = new int[200_000];
+                ActivationCountBuffer = new long[200_000];
                 CyclesBuffer = new long[200_000];
                 pointer = 1;
         }
